Fix: cast size_t to unsigned long for 32bit plattforms

+++ src/lib/sds_extras.c
@@ -206,7 +206,7 @@ int sds_getline(sds *s, FILE *fp, size_t max) {
             i++;
         }
         else {
-            MYMPD_LOG_ERROR("Line is too long, max length is %lu", max);
+            MYMPD_LOG_ERROR("Line is too long, max length is %lu", (unsigned long)max);
             return -2;
         }
     }
@@ -225,7 +225,7 @@ int sds_getfile(sds *s, FILE *fp, size_t max) {
         int c = fgetc(fp);
         if (c == EOF) {
             sdstrim(*s, "\r \t\n");
-            MYMPD_LOG_DEBUG("Read %lu bytes from file", sdslen(*s));
+            MYMPD_LOG_DEBUG("Read %lu bytes from file", (unsigned long)sdslen(*s));
             if (sdslen(*s) > 0) {
                 return 0;
             }
@@ -236,7 +236,7 @@ int sds_getfile(sds *s, FILE *fp, size_t max) {
             i++;
         }
         else {
-            MYMPD_LOG_ERROR("File is too long, max length is %lu", max);
+            MYMPD_LOG_ERROR("File is too long, max length is %lu", (unsigned long)max);
             return -2;
         }
     }
