+++ src/web_server.c
@@ -250,7 +250,7 @@ static void send_api_response(struct mg_mgr *mgr, struct t_work_result *response
     struct mg_connection *nc = mgr->conns;
     while (nc != NULL) {
         if ((int)nc->is_websocket == 0 && nc->id == (long unsigned)response->conn_id) {
-            MYMPD_LOG_DEBUG("Sending response to conn_id %lu (length: %lu): %s", nc->id, sdslen(response->data), response->data);
+            MYMPD_LOG_DEBUG("Sending response to conn_id %lu (length: %lu): %s", nc->id, (unsigned long)sdslen(response->data), response->data);
             if (response->cmd_id == INTERNAL_API_ALBUMART) {
                 webserver_albumart_send(nc, response->data, response->binary);
             }
@@ -381,14 +381,14 @@ static void ev_handler(struct mg_connection *nc, int ev, void *ev_data, void *fn
             }
             //check uri length
             if (hm->uri.len > URI_LENGTH_MAX) {
-                MYMPD_LOG_ERROR("Uri is too long, length is %lu, maximum length is %d", hm->uri.len, URI_LENGTH_MAX);
+                MYMPD_LOG_ERROR("Uri is too long, length is %lu, maximum length is %d", (unsigned long)hm->uri.len, URI_LENGTH_MAX);
                 nc->is_closing = 1;
                 return;
             }
 
             //check post requests length
             if (nc->label[1] == 'P' && (hm->body.len == 0 || hm->body.len > BODY_SIZE_MAX)) {
-                MYMPD_LOG_ERROR("POST request with body of size %lu is out of bounds", hm->body.len);
+                MYMPD_LOG_ERROR("POST request with body of size %lu is out of bounds", (unsigned long)hm->body.len);
                 nc->is_closing = 1;
                 return;
             }
